Finalized Omens ability, and Refined the CooldownText
++ Added UpdateCooldownText method in UIManager

++ Added Setters/Getter methods for clarity

++ Tweaked settings for testing, etc

+++ Assets/Scripts/Player/P-BulletManager.cs
@@ -8,7 +8,7 @@ public class PBulletManager : MonoBehaviour
 {
 	[Header("Player Bullet Settings")]
 	[SerializeField] private float _playerBulletDamage = 15f;
-	[SerializeField] private float _bulletMagazineCount = 140f;
+	[SerializeField] private int _bulletMagazineCount = 140;
 
 	[Header("Player Bullet Sprite Components")]
 	[SerializeField] private SpriteRenderer _thinBulletSprite;
@@ -21,7 +21,7 @@ public class PBulletManager : MonoBehaviour
 	{
 		//Enables Regular Bullet Sprite if Thruster Type is "Normal" or "Double"
 		this._regularBulletSprite.enabled =
-			(currentType == PWeaponManager.ThrusterType.Normal || currentType == PWeaponManager.ThrusterType.Double);
+		(currentType == PWeaponManager.ThrusterType.Normal || currentType == PWeaponManager.ThrusterType.Double);
 
 		//Enables Regular Bullet Capsule Collider if Thruster Type is not Wide ("Normal", "Double", or "Thin")
 		this._regularBulletCapCollider.enabled = (currentType != PWeaponManager.ThrusterType.Wide);
@@ -35,17 +35,17 @@ public class PBulletManager : MonoBehaviour
 	}
 
 	#region Setters & Getters
-	public void SetMagazineCount(float newBulletMagCount)
+	public void SetMagazineCount(int newBulletMagCount)
 	{
 		this._bulletMagazineCount = newBulletMagCount;
 	}
 
-	public float DecrementMagazineCount()
+	public int DecrementMagazineCount()
 	{
 		return --this._bulletMagazineCount;
 	}
 
-	public float GetCurrentMagazineCount()
+	public int GetCurrentMagazineCount()
 	{
 		return this._bulletMagazineCount;
 	}
